dashboard/pyproject.toml
[project]
name = "cs-market-tracker"
version = "0.1.0"
description = "i want the m9"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.13.1",
    "cloudscraper>=1.2.71",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]

[tool.uv]
cache-dir = ".uv-cache"
